chore: clean up Markdown.tsx code
